Split gameplay rejection feedback tests
